[TYPES] Bug fixes.
