+++ modules/api/src/main/scala/org/platanios/tensorflow/api/io/CheckpointReader.scala
@@ -65,10 +65,10 @@ class CheckpointReader private[CheckpointReader] (
     * @throws UnavailableException If this checkpoint reader object has already been disposed.
     */
   @throws[UnavailableException]
-  def getTensor[D <: DataType](name: String): Option[Tensor[D]] = {
+  def getTensor[T](name: String): Option[Tensor[T]] = {
     if (nativeHandle == 0)
       throw UnavailableException("This checkpoint reader has already been disposed.")
-    Option(NativeCheckpointReader.getTensor(nativeHandle, name)).map(Tensor.fromNativeHandle[D])
+    Option(NativeCheckpointReader.getTensor(nativeHandle, name)).map(Tensor.fromNativeHandle[T])
   }
 
   /** Returns a map from variable name to shape, for all variables containing in this checkpoint. */
@@ -78,9 +78,9 @@ class CheckpointReader private[CheckpointReader] (
   }
 
   /** Returns a map from variable name to data type, for all variables containing in this checkpoint. */
-  def variableDataTypes: Map[String, DataType] = {
+  def variableDataTypes: Map[String, DataType[_]] = {
     val types = NativeCheckpointReader.variableDataTypes(nativeHandle)
-    types.variables.zip(types.dataTypes.map(DataType.fromCValue[DataType])).toMap
+    types.variables.zip(types.dataTypes.map(DataType.fromCValue[DataType[_]])).toMap
   }
 }
